+++ src/NavFilesInFolder.h
@@ -1,0 +1,6 @@
+/* Copyright 2022 the SumatraPDF project authors (see AUTHORS file).
+   License: GPLv3 */
+
+struct MainWindow;
+
+void ShowNavFilesInFolder(MainWindow* win);